+++ DataAccess/Models/Permission.cs
@@ -8,8 +8,8 @@ namespace DataAccess.Models
 {
     public class Permission
     {
-        public long PermissionId { get; set; }
-        public long RoleId { get; set; }
+        public int PermissionId { get; set; }
+        public int RoleId { get; set; }
         public virtual Role Role { get; set; }
     }
 }
